+++ test/pages/habits/habits_tab_page_test.dart
@@ -1,6 +1,7 @@
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:lotti/blocs/habits/habits_cubit.dart';
+import 'package:lotti/blocs/habits/habits_state.dart';
 import 'package:lotti/classes/journal_entities.dart';
 import 'package:lotti/database/database.dart';
 import 'package:lotti/get_it.dart';
@@ -63,11 +64,13 @@ void main() {
     tearDown(getIt.reset);
 
     testWidgets('habits page is rendered', (tester) async {
+      final cubit = HabitsCubit()..setDisplayFilter(HabitDisplayFilter.all);
+
       await tester.pumpWidget(
         makeTestableWidgetWithScaffold(
           BlocProvider<HabitsCubit>(
             lazy: false,
-            create: (_) => HabitsCubit(),
+            create: (_) => cubit,
             child: const HabitsTabPage(),
           ),
         ),