+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added:
+- Segmented control for filtering which habits are shown (due, later today, complete, all)
+
 ### Changed:
 - Selectable habit time spans
 - More obvious habit completion state with strike-trough text and subtle opacity